update tests to use SdkGlobalStylesWrapper
